server/requirements.txt
fastapi == 0.89.1
graphene == 3.2.1
sqlalchemy == 2.0.1
uvicorn == 0.20.0
faker == 16.6.1
strawberry-graphql[fastapi] == 0.155.4
python-jose[cryptography] == 3.3.0
passlib[bcrypt] == 1.7.4
python-multipart == 0.0.5
kink == 0.6.5